+++ app.js
@@ -3,7 +3,7 @@ import dotenv from 'dotenv'
 import connectDB from './task-management/config/db.js'
 import authRoutes from './task-management/routes/authRoutes.js'
 import errorHandler from './task-management/middlewares/errors/errorHandler.js'
-import corsMiddleware from './task-management/security/cors/corsMiddleware.js'
+import corsMiddleware from './task-management/security/cors/middlewares/corsMiddleware.js'
 import handlePreflight from './task-management/security/preflight/handlePreflight.js'
 import applyMiddlewares from './task-management/middlewares/express/expressMiddleware.js'
 import { sendEmail } from './task-management/controllers/emails/emailController.js'
